Make SVG descendants opaque to event handlers
If an event like 'click' is bound to a parent of a SVG and the user
clicks on a descendant of the SVG, e.target will be set to that
descendant. This is generally unexpected and naïve code might think
e.target is always the bound element which is not true in this case.